webwatches/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
cssutils==2.11.1
Django==5.1
django-allauth==65.1.0
django-crispy-forms==2.3
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.8
joblib==1.4.2
lxml==5.3.0
more-itertools==10.5.0
multidict==6.0.5
numpy==2.1.2
pandas==2.2.3
pillow==11.0.0
plyer==2.1.0
premailer==3.10.0
pydivert==2.1.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scapy==2.6.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
stripe==11.2.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==5.0.3
yagmail==0.15.293
yarl==1.9.4
